Restructure npm normalize tests
This restructions the npm normalize tests, putting the different
QUnit.modules into their own test files to make it easier to add new
tests.

Also adds a breaking test for #1208
